scanservjs-rag/app/requirements.txt
streamlit>=1.32
chromadb>=0.4.24
anthropic>=0.25
pytesseract>=0.3.10
pdf2image>=1.16
Pillow>=10.0
httpx>=0.26
python-magic>=0.4.27
watchdog>=3.0
pdfminer.six>=20221105
